// models/CheckIfMemberResponse.ts
/**
 * Sendbird Platform SDK
 * Sendbird Platform API SDK  [https://sendbird.com/docs/chat/v3/platform-api/getting-started/prepare-to-use-api](https://sendbird.com/docs/chat/v3/platform-api/getting-started/prepare-to-use-api)  Contact Support:   Name: Sendbird   Email: [support@sendbird.com](https://mailto:support@sendbird.com)
 *
 * OpenAPI spec version: 1.0.0
 * Contact: support@sendbird.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class CheckIfMemberResponse {
    'isMember'?: boolean;
    'state'?: CheckIfMemberResponseStateEnum;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "isMember",
            "baseName": "is_member",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "state",
            "baseName": "state",
            "type": "CheckIfMemberResponseStateEnum",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return CheckIfMemberResponse.attributeTypeMap;
    }

    public constructor() {
    }
}


export type CheckIfMemberResponseStateEnum = "invited" | "joined" | "none" ;